shell_funcs

+++ labs_god/lab4/shell_sort/shell_funcs.c
@@ -1,6 +1,6 @@
 // Сортировка вставками, но только элементов, отстающих друг от друга на величину gap
 // Если gap = 1, то это обычная сортировка вставками.
-void insertionSort(int array[], int size, int gap, int* compares, int *swaps)
+void insertionSort(int array[], int size, int gap, long long* compares, long long *swaps)
 {
     for (int i = gap; i < size; ++i){
         int tmp = array[i];
@@ -19,7 +19,7 @@ void insertionSort(int array[], int size, int gap, int* compares, int *swaps)
 }
 
 // Сортировка Шелла
-int* shellSort(int array[], int size, int *compares, int *swaps)
+int* shellSort(int array[], int size, long long *compares, long long *swaps)
 {
     for (int gap = size / 2; gap > 0; gap /= 2){
         insertionSort(array, size, gap, compares, swaps);
